feat: wire up gatherer data pipeline and fix runtime issues
  Connect all gatherer components (Connection Manager, Message Router,
  Writers) to enable real-time data collection from Kalshi WebSocket API.

  Pipeline wiring:
  - Wire up Connection Manager → Message Router → Writers in main.go
  - Fix startup order: start Router and Writers BEFORE Connection Manager
    to prevent buffer overflow during initial market subscription
  - Add proper shutdown handling with deferred cleanup

  Buffer sizing for 300K+ markets:
  - Increase per-client buffer: 1,000 → 100,000 messages
  - Increase central manager buffer: 10,000 → 1,000,000 messages
  - Router uses GrowableBuffer that auto-resizes at 70% capacity

  JSON parsing fixes:
  - Add FlexInt64 type to handle Kalshi timestamps sent as either
    strings ("1234567890") or numbers (1234567890)
  - Update orderbookDeltaWire, tradeWire, tickerWire to use FlexInt64

  Database fixes:
  - Fix ON CONFLICT column order to match primary key constraints:
    - orderbook_deltas: (exchange_ts, ticker, price, side)
    - orderbook_snapshots: (snapshot_ts, ticker, source)
  - Add unix_now_microseconds() function for TimescaleDB retention
    policies with BIGINT time columns

  Local development:
  - Add configs/local/gatherer.yaml for local testing

+++ internal/router/router.go
@@ -298,7 +298,7 @@ func (r *router) parseOrderbookDelta(raw connection.RawMessage) (OrderbookMsg, e
 		PriceDollars: wire.Msg.PriceDollars,
 		Delta:        wire.Msg.Delta,
 		Side:         wire.Msg.Side,
-		ExchangeTs:   wire.Msg.Ts * 1_000_000, // seconds → microseconds
+		ExchangeTs:   int64(wire.Msg.Ts) * 1_000_000, // seconds → microseconds
 	}, nil
 }
 
@@ -318,7 +318,7 @@ func (r *router) parseTrade(raw connection.RawMessage) (TradeMsg, error) {
 		TakerSide:       wire.Msg.TakerSide,
 		SID:             wire.SID,
 		Seq:             wire.Seq,
-		ExchangeTs:      wire.Msg.Ts * 1_000_000,
+		ExchangeTs:      int64(wire.Msg.Ts) * 1_000_000,
 		ReceivedAt:      raw.ReceivedAt,
 		SeqGap:          raw.SeqGap,
 		GapSize:         raw.GapSize,
@@ -343,7 +343,7 @@ func (r *router) parseTicker(raw connection.RawMessage) (TickerMsg, error) {
 		DollarVolume:       wire.Msg.DollarVolume,
 		DollarOpenInterest: wire.Msg.DollarOpenInterest,
 		SID:                wire.SID,
-		ExchangeTs:         wire.Msg.Ts * 1_000_000,
+		ExchangeTs:         int64(wire.Msg.Ts) * 1_000_000,
 		ReceivedAt:         raw.ReceivedAt,
 		// Note: ticker has no Seq field
 	}, nil
